Fix to output text:
The fields were always separated by comma, but floats were printed using the decimal separator from regional settings.
This means that the file format was broken when running in system using comma as decimal separator.
From this commit on, fields are separated using the list separator specified in the regional settings (as Excel do when generating CSV files).